cmprskWeb-package(cmprskWeb)	R Documentation

Competing Risks Models for Web Data

Description

Provides various functions for competing risks analyses on Web data.

Details

Package: 	 cmprskWeb

Type: 	 Package

Version: 	 0.1.1

Date: 	 2008-04-07

License: 	 GPL

Author(s)

Patrick Mair

References

Mair, P. and Hudec, M. (2008). Competing Risks Approaches for Modeling Dwell Time Based Page Transitions in Web Usage Mining. Under editorial review in Journal of Data Mining and Knowledge Discovery.

See Also

cmprsk ~~

Examples


data(webshopms)
library("cmprsk")
msres <- mscuminc(webshopms, cutpoint = 1000, censoring = TRUE)
msres
plot(msres, from = 5.2)

[Package cmprskWeb version 0.1-1 Index]