mscuminc(cmprskWeb)	R Documentation

Tests on the cumulative incidence function

Description

Computes the cumulative incidence function (CIF) and performs tests of subgroup equality. It is a wrapper function for cuminc() in the package cmprsk.

Usage


mscuminc(msdata, cutpoint = NULL, censoring = TRUE)

Arguments

msdata 	 Data prepared with msprep()

cutpoint 	 Integer value which denotes the lower dwell time bound

censoring 	 If TRUE dwell times larger than cutpoint will be regarded as censored, if FALSE eliminated.

Value

msdata 	 Input data after cutpoint treatment

msres 	 List with testing results for cumulative incidence function (from cuminc) for each page transition

See Also

msprep

Examples


data(webshopms)
library("cmprsk")
msres <- mscuminc(webshopms, cutpoint = 1000, censoring = TRUE)
msres
plot(msres, from = 5.2)

[Package cmprskWeb version 0.1-0 Index]