webshopms(cmprskWeb)	R Documentation

Webshop data for multistage competing risks modeling

Description

This data set for multistage computations is prepared with the function msprep()

Usage

data(webshopms)

Format

A data frame with 19836 observations on the following 7 variables.

session

Session ID

user

User ID

from

Starting page (see details)

to

Landing page

time

Dwell time on starting page

nvisit

visit number within clickstream

buy

buyer vs. non-buyer

Details

The starting pages are prepared in terms of the number of visits within a clickstream, e.g. "1.4" denotes the transitions after the 4th visit of page 1.

Examples


data(webshopms)
## maybe str(webshopms) ; plot(webshopms) ...

[Package cmprskWeb version 0.1-0 Index]